travel_rag_agent/requirements.txt
google-adk[a2a]==1.25.0
python-dotenv
pyyaml
google-cloud-aiplatform
